requirements.txt
mitmproxy>=10.0
python-dotenv>=1.0
requests>=2.28
